+++ src/org/starexec/data/to/CommunityRequest.java
@@ -8,37 +8,37 @@ import java.sql.Timestamp;
  * @author Todd Elvers & Tyler Jensen
  */
 public class CommunityRequest {
-	private Long communityId;
+	private int communityId;
 	private String message;
 	private String code;
 	private Timestamp createDate;
-	private long userId;
+	private int userId;
 	
 	/**
 	 * @return the user_id of the user who created the request
 	 */
-	public long getUserId(){
+	public int getUserId(){
 		return this.userId;
 	}
 	
 	/**
 	 * @param userId the userId of the user who created this invite
 	 */
-	public void setUserId(long userId){
+	public void setUserId(int userId){
 		this.userId = userId;
 	}
 	
 	/**
 	 * @return the id of the community this request is for
 	 */
-	public long getCommunityId() {
+	public int getCommunityId() {
 		return this.communityId;
 	}
 
 	/**
 	 * @param communityId the id of the community to set for this request
 	 */
-	public void setCommunityId(long communityId) {
+	public void setCommunityId(int communityId) {
 		this.communityId = communityId;
 	}
